fix:Ajustando nomeclaturas e também refatorando os teste unitarios e de integração
