+++ src/main/java/com/api/apibackend/modules/Cart/Application/DTOs/ProductCheckQuantity.java
@@ -1,5 +1,3 @@
-package com.api.apibackend.modules.Cart.Application.DTOs;
-
 /**
  * ----------------------------------------------------------------------------
  * Autor: Kaue de Matos
@@ -7,6 +5,8 @@ package com.api.apibackend.modules.Cart.Application.DTOs;
  * Propriedade da Empresa: Todos os direitos reservados
  * ----------------------------------------------------------------------------
  */
+
+package com.api.apibackend.modules.Cart.Application.DTOs;
 
 import java.util.List;
