+++ src/main/java/com/api/apibackend/modules/Cart/Domain/service/FraudDetectorOrder.java
@@ -1,5 +1,3 @@
-package com.api.apibackend.modules.Cart.Domain.service;
-
 /**
  * ----------------------------------------------------------------------------
  * Autor: Kaue de Matos
@@ -7,10 +5,9 @@ package com.api.apibackend.modules.Cart.Domain.service;
  * Propriedade da Empresa: Todos os direitos reservados
  * ----------------------------------------------------------------------------
  */
+package com.api.apibackend.modules.Cart.Domain.service;
 
 import org.springframework.stereotype.Service;
 
 @Service
-public class FraudDetectorOrder {
-    
-}
+public class FraudDetectorOrder { }
