+++ src/main/java/com/api/apibackend/modules/Midia/Application/useCases/CreatedMidia/MidiaController.java
@@ -6,10 +6,8 @@
  * ----------------------------------------------------------------------------
  */
 
-package com.api.apibackend.modules.Supplier.Domain.exception;
+package com.api.apibackend.modules.Midia.Application.useCases.CreatedMidia;
 
-public class ErrorEmptySupplier extends Exception {
-    public ErrorEmptySupplier(String message) {
-        super(message);
-    }
+public class MidiaController {
+
 }
