--- src/main/java/com/api/apibackend/modules/Midia/Domain/helpers/MidiaModelMapper.java
@@ -1,32 +1,0 @@
-/**
- * ----------------------------------------------------------------------------
- * Autor: Kaue de Matos
- * Empresa: Nova Software
- * Propriedade da Empresa: Todos os direitos reservados
- * ----------------------------------------------------------------------------
- */
-
-package com.api.apibackend.modules.Midia.Domain.helpers;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import com.api.apibackend.modules.Midia.Application.DTOs.MidiaDTO;
-import com.api.apibackend.modules.Midia.infra.persistence.entity.MidiaEntity;
-
-public class MidiaModelMapper {
-    private ModelMapper modelMapper;
-
-    @Autowired
-    public MidiaModelMapper(ModelMapper modelMapper) {
-        this.modelMapper = modelMapper;
-    }
-
-    public MidiaEntity toMidiaDTOFromMidiaEntity(MidiaDTO midiaDTO) {
-        return modelMapper.map(midiaDTO, MidiaEntity.class);
-    }
-
-    public MidiaDTO toMidiaEntityFromMidiaRequest(MidiaEntity requestMidiaEntity) {
-        return modelMapper.map(requestMidiaEntity, MidiaDTO.class);
-    }
-}
