+++ src/main/java/com/api/apibackend/modules/Midia/infra/persistence/repository/MidiaRepository.java
@@ -6,23 +6,19 @@
  * ----------------------------------------------------------------------------
  */
 
-package com.api.apibackend.modules.Midia.infra.repository;
+package com.api.apibackend.modules.Midia.infra.persistence.repository;
 
-import java.util.List;
-
+import com.api.apibackend.modules.Midia.infra.persistence.entity.MidiaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.api.apibackend.modules.Midia.infra.persistence.entity.MidiaEntity;
+import java.util.List;
 
 @Repository
 public interface MidiaRepository extends JpaRepository<MidiaEntity, Long> {
     
    List<MidiaEntity> findByCategory(String category);
-
-   @Query("SELECT COUNT(m) FROM MidiaEntity m WHERE m.category = 'Salads'")
-   Long countByCategorySalads();
 
    List<MidiaEntity> findAllByCategoryIn(List<String> categories);
 }
