+++ src/main/java/com/api/apibackend/modules/Product/Domain/service/ProductConventionalService.java
@@ -13,6 +13,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
+import com.api.apibackend.modules.Midia.infra.persistence.repository.MidiaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,7 +21,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.api.apibackend.modules.Midia.infra.persistence.entity.MidiaEntity;
-import com.api.apibackend.modules.Midia.infra.repository.MidiaRepository;
 import com.api.apibackend.modules.Product.Application.DTOs.ProductDTO;
 import com.api.apibackend.modules.Product.Application.DTOs.ResponseMessageDTO;
 import com.api.apibackend.modules.Product.Application.component.ProductComponentAdd;
@@ -31,7 +31,7 @@ import com.api.apibackend.modules.ProductCategory.infra.persistence.entity.Produ
 import com.api.apibackend.modules.ProductCategory.infra.persistence.repository.ProductCategoryRepository;
 import com.api.apibackend.modules.Stock.Infra.persistence.entity.StockEntity;
 import com.api.apibackend.modules.Stock.Infra.persistence.repository.StockRepository;
-import com.api.apibackend.modules.Supplier.Infra.entity.SupplierEntity;
+import com.api.apibackend.modules.Supplier.Infra.persistence.entity.SupplierEntity;
 import com.api.apibackend.modules.Unity.infra.persistence.entity.UnityEntity;
 import com.api.apibackend.modules.Unity.infra.persistence.repository.UnityRepository;
