+++ src/main/java/com/api/apibackend/modules/Supplier/Application/repository/ISupplierService.java
@@ -1,5 +1,3 @@
-package com.api.apibackend.modules.Supplier.Application.repository;
-
 /**
  * ----------------------------------------------------------------------------
  * Autor: Kaue de Matos
@@ -8,6 +6,8 @@ package com.api.apibackend.modules.Supplier.Application.repository;
  * ----------------------------------------------------------------------------
  */
 
+package com.api.apibackend.modules.Supplier.Application.repository;
+
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import com.api.apibackend.modules.Supplier.Application.DTOs.ResponseMessageDTO;
 import com.api.apibackend.modules.Supplier.Application.DTOs.SupplierDTO;
-import com.api.apibackend.modules.Supplier.Infra.entity.SupplierEntity;
+import com.api.apibackend.modules.Supplier.Infra.persistence.entity.SupplierEntity;
 
 public interface ISupplierService {
     public List<SupplierEntity> listSupplier();
