+++ src/main/java/com/api/apibackend/modules/Supplier/Application/useCases/SupplierDelete/SupplierDeleteController.java
@@ -1,3 +1,11 @@
+/**
+ * ----------------------------------------------------------------------------
+ * Autor: Kaue de Matos
+ * Empresa: Nova Software
+ * Propriedade da Empresa: Todos os direitos reservados
+ * ----------------------------------------------------------------------------
+ */
+
 package com.api.apibackend.modules.Supplier.Application.useCases.SupplierDelete;
 
 import org.springframework.beans.factory.annotation.Autowired;
