+++ src/main/java/com/api/apibackend/shared/helpers/ModelMappersConvertion.java
@@ -2,13 +2,15 @@ package com.api.apibackend.shared.helpers;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
+@Component
 public class ModelMappersConvertion<T, U> {
     private ModelMapper modelMapper;
 
     @Autowired
     public ModelMappersConvertion(ModelMapper modelMapper) {
-        this.modelMapper = modelMapper;
+        this.modelMapper = new ModelMapper();
     }
 
     public U toDTOFromEntity(T source, Class<U> destinationType) {
