+++ src/test/java/com/api/apibackend/modules/Auth/Infra/repository/IMemory/IUserRepository.java
@@ -1,3 +1,10 @@
+/**
+ * ----------------------------------------------------------------------------
+ * Autor: Kaue de Matos
+ * Empresa: Nova Software
+ * Propriedade da Empresa: Todos os direitos reservados
+ * ----------------------------------------------------------------------------
+ */
 package com.api.apibackend.modules.Auth.Infra.repository.IMemory;
 
 import com.api.apibackend.modules.Auth.Application.DTOs.UserDTO;
